test microservice vs monolithic

+++ docker/api-gateway/run.sh
@@ -1,13 +1,17 @@
 #!/bin/bash
 
 # API Gateway
+cd /var/www/
+declare -a arr=("ApiGateway" "TransactionsService" "ProductsService" "QRBenefit-Monolithic")
 
-declare -a arr=("ApiGateway" "OrdersService" "ProductsApi" "QRBenefitService")
-
-for i in "${arr[@]}"
+for project in "${arr[@]}"
 do
-   cd /var/www/$i
-   # Permission
+   cd $project
+    if [ ! -d "$i" ]; then
+        break
+    fi
+
+   # Update permission directory
     if [ "$(stat -c "%U" storage/logs/)" != "www-data" ] || [ "$(stat -c "%G" storage/logs/)" != "www-data" ]; then
         rm -rf sstorage/logs/*
         chown -R www-data:www-data storage/logs/
